clean directories
